requirements.txt
langchain
streamlit
python-dotenv
openai
st-chat-message
pandas
tiktoken
